BEnemyのステートマシン化2
FollowEnemyのスクリプトを参考に実装
